embeddings/embedder/embedder.py
from sentence_transformers import SentenceTransformer
import torch

device = "cuda" if torch.cuda.is_available() else "cpu"
embedder = SentenceTransformer('all-MiniLM-L6-v2', device=device)


def embed_text(text):
    return embedder.encode(text).tolist()