embeddings/requirements.txt
psycopg2-binary
requests
transformers
sentence-transformers